RAG_Version 1.3/requirement.txt
# requirements.txt
requests
numpy
scikit-learn
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-core
langchain-huggingface
sentence-transformers
pinecone-client
langchain-pinecone
GitPython
